coala: Add file caching
In this commit, coalib.misc.Caching is introduced which contains two
functions: `update_last_coala_run_time` and `get_last_coala_run_time`
which can be used to update and retrieve the last time coala was run.

This will be useful to determine which files are new or changed and
run coala only on those files to improve coala run time speed.